+++ botfront/cypress/support/response.commands.js
@@ -73,3 +73,17 @@ Cypress.Commands.add('addButtonOrSetPayload', (title, payload, oldTitleOrPositio
         cy.dataCy('save-button').click({ force: true });
     }
 });
+
+Cypress.Commands.add('setQuickReplyContent', (text, title, intent, index = 0) => {
+    cy.dataCy('response-editor').find('[data-cy=button_title]').eq(index).click({ force: true });
+    cy.dataCy('enter-button-title').find('input').type(title);
+    cy.dataCy('intent-label').should('exist').click();
+    cy.dataCy('intent-dropdown').find('input').type(`${intent}{enter}`);
+    cy.dataCy('save-button').click();
+    cy.dataCy('response-editor').find('[data-cy=bot-response-input]')
+        .find('textarea')
+        .eq(index)
+        .clear()
+        .type(text)
+        .blur();
+});