+++ botfront/imports/ui/components/nlu/activity/ActivityCommandBar.jsx
@@ -69,6 +69,7 @@ const ActivityCommandBar = React.forwardRef((props, ref) => {
                     allowEditing
                     onChange={intent => onSetIntent(selection, intent)}
                     onClose={onCloseIntentPopup}
+                    enableReset
                 />
                 <Popup
                     size='mini'