+++ botfront/imports/ui/components/templates/templates-list/BotResponseEditor.jsx
@@ -23,6 +23,7 @@ import {
     addResponseLanguage,
     getDefaultTemplateFromSequence,
     addContentType,
+    modifyResponseType,
 } from '../../../../lib/botResponse.utils';
 import { clearTypenameField } from '../../../../lib/client.safe.utils';
 import { Loading } from '../../utils/Utils';
@@ -97,7 +98,7 @@ const BotResponseEditor = (props) => {
         );
     };
 
-    const updateResponse = (updatedResponse, callback) => {
+    const updateResponse = (updatedResponse, callback = () => {}) => {
         updateBotResponse({
             variables: {
                 projectId, _id: updatedResponse._id, response: clearTypenameField(updatedResponse),
@@ -180,6 +181,15 @@ const BotResponseEditor = (props) => {
         upsertResponse(name, updatedSequence, index);
     };
 
+    const handleChangePayloadType = (newType) => {
+        const updatedBotResponse = modifyResponseType(newBotResponse, newType);
+        if (isNew || !newBotResponse._id) {
+            setNewBotResponse(updatedBotResponse);
+            return;
+        }
+        updateResponse(updatedBotResponse);
+    };
+
     const getActiveSequence = () => {
         const activeValue = newBotResponse.values && newBotResponse.values.find(({ lang }) => lang === language);
         if (!activeValue) {
@@ -241,6 +251,7 @@ const BotResponseEditor = (props) => {
                     sequence={activeSequence}
                     onChange={handleSequenceChange}
                     onDeleteVariation={handleDeleteVariation}
+                    onChangePayloadType={handleChangePayloadType}
                     name={name}
                     editable={editable}
                 />